Moved EndToEnd tests to a separate package
